Support storage path permissions disablement
