+++ api/profiles/release/application.properties
@@ -114,7 +114,7 @@ data.storage.nfs.root.mount.point=
 # Mount options for NFS
 data.storage.nfs.options.rsize=1048576
 data.storage.nfs.options.wsize=1048576
-data.storage.enable.security=${CP_DATA_STORAGE_ENABLE_SECURITY:true}
+data.storage.path.permissions.enabled=${CP_DATA_STORAGE_PATH_PERMISSIONS_ENABLED:true}
 
 data.storage.nfs.quota.poll=${CP_API_NFS_QUOTA_POLL:60000}
 data.storage.nfs.quota.metadata.key=${CP_API_NFS_QUOTA_METADATA_KEY:fs_notifications}
