+++ api/src/test/resources/test-application.properties
@@ -144,7 +144,7 @@ data.storage.nfs.root.mount.point=pipeline-test
 data.storage.nfs.options.rsize=1048576
 data.storage.nfs.options.wsize=1048576
 data.storage.nfs.events.enable.sync=false
-data.storage.enable.security=false
+data.storage.path.permissions.enabled=false
 
 cluster.enable.autoscaling=false
