+++ deploy/docker/cp-api-srv/config/application.properties
@@ -167,7 +167,7 @@ luigi.graph.script=/usr/lib/python2.7/site-packages/scripts/deps_graph.py
 
 # File storages
 data.storage.nfs.root.mount.point=/opt/api/file-systems
-data.storage.enable.security=${CP_DATA_STORAGE_ENABLE_SECURITY:true}
+data.storage.path.permissions.enabled=${CP_DATA_STORAGE_PATH_PERMISSIONS_ENABLED:true}
 
 #Billing API
 billing.index.common.prefix=cp-billing
